:recycle: Move VoteUrlProcessorV2 component into Voting folder
... and remove 'V2' from the name
